React 16 Updates
Upgrade peer dependency to react 16, and use a forwardRef to set a ref
to the wrapped component, rather than a regular ref.  This helps it work
with function components.

Bump react dependency to 16.3 to guarantee forwardRef exists.  Bump
Enzyme to compatible version that works with React 16.  Fix failing
test. Bump package version to 2.0.0 since this a breaking change from
how refs previously worked.